make the fault transition be "fade"
The transition that was in the original pinpoint, with the presentation running
down the right hand side and text flying in and out from there was getting old,
and anyways has had a regression with clutter changing the distance to the far
clipping plane. Fades are simple and nice - and like black will continue being
fashionable.