pinpoint: Remove all those weird warning about opening "./NULL"
It was because "NULL" was the default bg which was treated as an image
file. The Clutter code was relying on that failing to create an empty
texture as the backround.

Like, really?

Just handle the NONE background explictely by create a dummy rectangle
of the same color then the stage. This is a way to keep the invariant of
always having a valid ClutterActor as the background.