clutter: Don't make X calls if the Clutter backend isn't X11
Previously pinpoint was using #ifdef's to determine whether to make
calls specific to the X11 backend of Clutter when fullscreening or
inhibiting the screensaver. Clutter now has run-time support for
multiple backends so this would break when trying to use a build of
Clutter that has both Wayland and X11 support. This patch makes it
check the class name of the Clutter backend before using X11 calls.
Clutter doesn't appear to have a better way to determine which backend
is being used.